Island: Rename repository file names to snake case
