+++ monkey/monkey_island/cc/repository/i_attack_repository.py
