+++ monkey/monkey_island/cc/repository/i_config_repository.py
