+++ monkey/monkey_island/cc/repository/i_log_repository.py
