+++ monkey/monkey_island/cc/repository/i_machine_repository.py
