+++ monkey/monkey_island/cc/repository/i_network_map_repository.py
