+++ monkey/monkey_island/cc/repository/i_simulation_repository.py
