+++ monkey/monkey_island/cc/repository/i_stolen_credentials_repository.py
