+++ monkey/monkey_island/cc/repository/i_telemetry_repository.py
